remove tests, start working on workflows/fieldmap/unwarp.py
these tests check that a list of hard-coded node names are effectively there. In practical terms I don't see any benefit on it.

If there is any way to test nipype workflows other than smoke tests, let's develop more that idea. I understand the appealing of having unit tests on workflows, but I don't think we have any good solution to the problem. We'll need to stick with smoke tests.