[HOTFIX] Sphinx 1.7.0
As suspected, this bugfix (sphinx-doc/sphinx#2557) to sphinx broke our builds, as mocked modules do not mock ancestors anymore. An updated mocks list (only for sphinx>=1.7.0). Version is checked with packaging.version to avoid distutils (also, setuptools depends on packaging so it was installed before declaring it as a dependency).

Fixes #973.