ENH: Stop using siemens2rads from old nipype workflows
A new interface is proposed, to be used in phase-difference based
workflows.

@AKSoo identified a potential issue with siemens2rads
(https://github.com/poldracklab/sdcflows/pull/30#discussion_r346640162),
although the re-centering did not seem to make a big difference.

This implementation of the rescaling uses percentiles to robustly
calculate the range of the input image (as recommended in the original
paper about FSL PRELUDE), and makes sure the output data type is
adequate (float32).