ENH: Deduplicating magnitude handling and fieldmap postprocessing workflows
This PR splits the magnitude processing and the fieldmap postprocessing
that is shared between phasediff (and phase1/2) and direct B0 mapping
approaches.

It also addresses some issues on the direct B0 mapping approach, where
the fieldmap was converted between units and subject to PRELUDE (which
is not correct as these fieldmaps are not phase-wrapped).

Adding a test case of SE fieldmaps would be great for a future PR.

Close #17