fix: revise total-readout-time, cleanup ees
This commit makes a deep revisiting of our total readout time implementation.
Most of the relevant details about this commit are included within the
documentation.

With thanks to @neurolabusc because this wouldn't have been possible without
his thorough work, to @parekhpravesh for the thread he opened at
Neurostars, to @Gilles86 for making us aware of these issues on the *fMRIPrep*
repo, and to @treanus for reporting #5.

Resolves: #5.