ENH: Disallow ``FieldmapEstimation`` objects with same sources
This PR implements a bi-directional hash-map (dictionary) instead of a
set to track fieldmap estimator IDs.

Instead, the registry can be queried by passing in the fieldmap
estimator ID, OR a tuple of path-sources (in which case it returns
the ID).

This way, a ``ValueError`` is raised when one tries to establish
a fieldmap estimation using the same set of images.

Resolves: #129.