ENH: Massage TOPUP's fieldcoeff files to be compatible with ours
TOUP-generated "fieldcoeff" files are just B-Spline fields, where the shape
of the field is fixated to be a decimated grid of the original image by a
factor of 2x and added 3 pixels on each dimension.
This is one root reason why TOPUP errors (FSL 6) or segfaults (FSL5), when the
input image has odd number of voxels along one or more directions.

These "fieldcoeff" are fixated to be zero-centered, and have "plumb" orientation
(as in, aligned with cardinal/imaging axes).
The q-form of these NIfTI files is always diagonal, with the decimation factors
set on the diagonal (and hence, the voxel zooms).
The origin of the q-form is set to the reference image's shape.

This interface modifies these coefficient files to be fully-fledged NIfTI images
aligned with the reference image.
Therefore, the s-form header of the coefficients file is updated to match that
of the reference file.
The s-form header is used because the imaging axes may be oblique.

The q-form retains the original header and is marked with code 0.