enh: optimize implementation using tensor-product of 1D vectors
This implementation hopes to be more optimal than the previous one.
Instead of calculating the tensor-product B-Spline weights of every
point w.r.t. every control point, it calculates the weights along
each axis and then calculates the tensor-product for the whole
grid.

In principle, the number of calls to ``np.piecewise`` has to have
dramatically dropped.
Memory utilization should be also more optimal, as there're only
one very short-lived and large array (before it is converted to
sparse matrix).