FIX: Make images "plumb" before running ANTs-SyN (and roll-back afterward)
Because ITK handles registration in "physical" coordinates, when the fixed
image's affine is not aligned with the array's axis (i.e., is not plumb or
aligned to the canonical axis) the transform is obtained with alignment to
one axis, but that is not aligned with the data.

To ensure the PE (phase-encoding) direction of the data array and the
corresponding physical axis are aligned, this PR compensates for potental
rotations.

Resolves #12.