FIX: Address memory issues of ``Coefficients2Warp``
This PR ends with the dual implementation of the evaluation of B-Spline
weights on the grid of the approximated map.
On the one hand, we had a ``grid_bspline_weights`` function, that used
vector operations, half-way optimized.
On the other, I later wrote ``bspline_weights`` as the former was untenable
by the approximation interface.

I vaguely remember having plans to unify them but left the code there untouched.

These changes should ease the memory issues of ``Coefficients2Warp`` (still hooked
with the first implementation).

Resolves: #194.