FIX: Normalize phase-encoding polarity of coefficients after TOPUP
It seems that TOPUP's coefficients polarity is referred to that of the
first EPI fed into topup.
This means that, if the polarity is reversed (e.g., ``"j-"``), then the
fieldmap will always be reversed.
This PR ensure the coefficients are "unreversed" if the polarity of the
reference is reverses.