FIX: Generation of RAS displacements fields from VSMs
After much thought, I have come to the conclusion that the VOX2RAS
affine must be applied in all cases, regardless of whether the dataset
is oblique or not.

After all, the VSM (voxel-shift-map) is no more than a regularly gridded
field of voxel coordinates.
The transformation between voxels and mm is biyective and univocal, and
formalized by the affine matrix of the image.

This PR simplifies the calculation, and theoretically should resolve
most of the issues we are experiencing when resampling data.

Resolves: #218.
Resolves: #236.

Related: nipreps/fmriprep#2210.