enh: I *believe* topup's coeffs should be flipped / sty: add comments
I have empirically tested this with Gilles' dataset (oriented LAS, PE
along *i*) and with @mgxd's dataset (oriented LAS too, but PE is *j*),
so this seems to be accurate.

I tested it by manually opening topup's and our fieldmap with Mango and
removing the affine's orientation (although switching it off should not
and does not alter the result), and checking that data are in the same
LR ordering.