enh(BSplineApprox): integrate downsampling when the input is high-res
Adds a new input ``zooms_min`` that will trigger the downsampling of the
input field map when it comes with a very high resolution (typically
this is the case for SDC SyN).

Although the input is resampled, the output field is interpolated at the
original resolution for debugging purposes.

Downsampling dramatically speeds up approximation.
Alternatively, a solution (for the SDC SyN) would be to calculate a
heavily dilated mask of the brain, and use it to limit the number of
voxels that are fit in regression.