fix/doc: revise docstrings, remove cache validation
CRITICAL change

``sdcflows/interfaces/tests/test_bspline.py::test_bsplines`` is a
parametric test where random coefficients are generated and then re-fit.
The previous resampler was not working.

However, fixing the resampling has made this test a point of concern. We
will need to revisit this test and make sure the bspline fitting is
acceptable.