fix: revise registration workflow
Since the affine between the target EPI and the fieldmap reference is
now applied "on-the-fly", the coefficients SHOULD NOT be reoriented
into the target space.

This also revises the test on the correction workflow, which also ran
registration (stacking potential problems by exercising several units)
so now the alignment matrix has been calculated offline and we can run
the test with and without it.

This commit requires the derivatives of the HCP dataset be up-to-date.