fix+enh: wrong transform creating NIfTI object + move to transform points convention
The wrong affine was assigned to the NIfTI object carrying the resampled
fieldmap, which did not fail if no transform between fieldmap and EPI
was given.

This PR also changes the naming of variables from BIDS' ``mode-image``
convention to ``mode-points`` convention, as the latter is a definition
closer to what we are actually doing (transforming points).