FIX: Mask fieldmap before fitting spline field
Previously, we fit a spline field to the within-mask portion of a
fieldmap, which could lead to large peaks outside the mask. Applying the
mask to the reconstructed field can produce discontinuities in
resampling. Instead of attempting to attenuate the peaks outside the
mask in a smooth way, we set the fit values outside the mask to zero,
and let the spline fit find a smooth field.

In practice, we've found that even fields with large peaks on the edge
of the mask are well fit by this method.